requirements/_base.txt
facebook_business==8.0.3
pyyaml==5.3.1
flake8
isort
Flask
flask-sqlalchemy
requests-oauthlib==1.3.0
rauth==0.7.3
